requirements.txt
streamlit
Pillowg